refactor processing multiple queries for filtering the df see alxndrkalinin/copairs@5d9b2b3f7a802d90c28aa69d61faf20c27b88793
